pyproject.toml
[project]
name = "llamaindex-fastapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "google-auth>=2.35.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.1",
    "llama-index>=0.11.16",
    "pydantic-settings>=2.5.2",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "llama-index-readers-google>=0.4.1",
]
